+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add an optional exit_code entry in ContainerInstance
 - Add OPAWDL
 ### Changed
+- Refactor OneDocker Runner to use granular exit codes
 ### Removed
 - Delete deprecated MPC entities in FBPCP (moved to FBPCS)
 - Replace invalid S3 bucket from OneDocker unit tests with mock path